added helper functions for isA, getTag fixed a warning on x64
git-svn-id: https://bugengine.googlecode.com/svn/trunk@471 f3700e90-1ba7-11de-b56b-c9554f25afe1

+++ src/engine/rtti/src/engine/classinfo.cc
@@ -116,7 +116,7 @@ bool ClassInfo::isA(weak<const ClassInfo> klass) const
     return false;
 }
 
-Value ClassInfo::operator()(Value* params, size_t nparams) const
+Value ClassInfo::operator()(Value* params, u32 nparams) const
 {
     return constructor(params, nparams);
 }
